+++ src/request/trips-request/trips-request.ts
@@ -220,7 +220,8 @@ export class TripRequest extends OJPBaseRequest {
       "car_sharing",
       "escooter_rental",
     ];
-    const isSharingMode = sharingModes.indexOf(transportMode) !== -1;    
+    const isSharingMode = sharingModes.indexOf(transportMode) !== -1;
+    const isWalking = transportMode === 'walk' || transportMode === 'foot';
 
     const nameNodeName = isOJPv2 ? 'Name' : 'LocationName';
 
@@ -267,7 +268,16 @@ export class TripRequest extends OJPBaseRequest {
         }
       }
 
-      if (!isMonomodal) {
+      if (isMonomodal) {
+        if (isOJPv2 && isWalking && isFrom) {
+          const transportOptionNode = endPointNode.ele(ojpPrefix + 'IndividualTransportOption');
+          
+          const personalModeNode = transportOptionNode.ele(ojpPrefix + 'ItModeAndModeOfOperation');
+          personalModeNode.ele(ojpPrefix + 'PersonalMode', 'foot');
+
+          transportOptionNode.ele(ojpPrefix + 'MaxDuration', 'PT60M');
+        }
+      } else {
         if (isOJPv2) {
           if (personalMode !== null) {
             (() => {
@@ -360,10 +370,7 @@ export class TripRequest extends OJPBaseRequest {
         }
       }
 
-      if (isSharingMode && isOJPv2) {
-        // NumberOfResults = 0 for sharing in OJP v2.0
-        paramsNode.ele(ojpPrefix + 'NumberOfResults', 0);
-      } else {
+      if (transportMode === 'public_transport') {
         if (this.numberOfResults !== null) {
           paramsNode.ele(ojpPrefix + 'NumberOfResults', this.numberOfResults);
         }
@@ -372,6 +379,13 @@ export class TripRequest extends OJPBaseRequest {
         }
         if (this.numberOfResultsAfter !== null) {
           paramsNode.ele(ojpPrefix + 'NumberOfResultsAfter', this.numberOfResultsAfter);
+        }
+      }
+
+      if (isOJPv2) {
+        if (isSharingMode || isWalking) {
+          // NumberOfResults = 0 for sharing / walking in OJP v2.0
+          paramsNode.ele(ojpPrefix + 'NumberOfResults', 0);
         }
       }
     }